# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Ultra Bank",
    "description": "The Ultra Bank apllication API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Bank",
      "description": "API for users in the system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],

  "paths": {
    "/account/all": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users information in the system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/account/findOne/{email}": {
      "parameters": [
        {
          "name": "email",
          "in": "path",
          "required": true,
          "description": "Email of user that we want information for",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Users"],
        "summary": "Get information for a specific user",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/account/findTransactions/{email}": {
      "parameters": [
        {
          "name": "email",
          "in": "path",
          "required": true,
          "description": "Email of user that we want to get transactions for",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Transactions"],
        "summary": "Get all transactions for a specific user in the system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Transaction"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "name": { "type": "string" },
        "email": { "type": "string" },
        "accountNumber": { "type": "string" },
        "balance": { "type": "integer" },
        "gender": { "type": "string" },
        "dob": { "type": "string" },
        "phoneNumber": { "type": "string" },
        "address": { "type": "string" }
      },
      "xml": { "name": "User" }
    },
    "Transaction": {
      "type": "object",
      "properties": {
        "email": { "type": "string" },
        "date": { "type": "string" },
        "type": { "type": "string" },
        "amount": { "type": "string" },
        "currentBalance": { "type": "integer" }
      },
      "xml": { "name": "User" }
    }
  }
}